+++ deepl-java/src/main/java/com/deepl/api/DeepLClientOptions.java
@@ -1,0 +1,7 @@
+// Copyright 2025 DeepL SE (https://www.deepl.com)
+// Use of this source code is governed by an MIT
+// license that can be found in the LICENSE file.
+package com.deepl.api;
+
+/** {@inheritDoc} */
+public class DeepLClientOptions extends TranslatorOptions {}